Add dwas_controller service and update connection settings
This commit introduces the dwas_controller to the docker-compose configuration, along with necessary dependancies and ports. Also, connection settings for the QUEUE_HOST have been updated and logging statements added in the process of connecting to RabbitMQ. The Dockerfile for the controller has been added and the Makefile has been updated with commands for building and rebuilding the Docker image.