Update temporary directory path and improve error logging
This commit updates the temporary directory path in the docker-compose and config files, increasing the flexibility of our program. Additionally, an 'Error' field was added to the 'Message' struct to improve error logging which would facilitate in debugging. Dockerfiles were also updated for 'release' setting for GIN_MODE in both 'dwas-controller' and 'dwas-worker'.