Add "QueueStageFinished" to Queue configuration
Introduced a new QueueStageFinished constant in ConfigurationConsts.go, and updated the corresponding config.yml and the QueueConfig struct in configuration.go to include this new queue stage. Made necessary changes in yml-configuration.go to fetch the value from the configuration file. This will allow tracking the finished state of queue stages.