Implement Docker container creation functionality in worker actions
A new file called docker-run.go has been added under the worker's actions. This file includes the creation of a Docker container, which is a new action in the worker's actions. The action takes an image property as input and outputs a docker-container-id property meant to hold the identifier of the created docker container. The interaction with the Docker server