Refactor pipeline parameters and add action-form view
Renamed the 'PipelineID' attribute in the 'Parameters' struct to 'StepID' for consistency. Added an action-form html view to present a form for each action. Also, initiated 'ParametersRepository' to fetch parameters for a given step ID from the database. Updated 'main.go' to include routes and handlers related to these changes.