Refactor pipeline creation and retrieval process
Reorganized the code to have a more streamlined and logical structure for creating and retrieving pipelines. This includes the separation of pipeline creation and pipeline retrieval into separate handlers as well as adding a pipeline link in the home menu. This also involved the adjustment of routing in the register modules to match these changes.