Add support for 'for-each' action in the pipeline
Updated the system to handle a new pipeline action called 'for-each'. This action allows us to iterate over collections. Changes included enabling sub-pipeline mappings and additional validations within the pipeline and stage structures. The 'for-each' action was also added in the map-executor and necessary adjustments in the pipeline parameters were made in the controller.http file.