Update ForeachBody type for improved loop sequence processing
The ForeachBody type has been restructured to allow for better management of for-each loops in stages. The SubPipeline field in the stage.go file has been substituted with ForeachBody, which provides more accurate sequencing of steps. Additionally, this revision includes the introduction of a 'parameters' map in the ForeachBody, which aids in array management.