Add main_test and corresponding Makefile rule
A new file, main_test.go, has been added in the tests package to handle the application's testing. In addition, a new rule 'test:' has been added to the Makefile, allowing the execution of the main_test via a simple 'go test'. The new test aims to ensure that the Docker compose setup is functioning as expected.