Add new functions 'addToQueue' and 'unmarshal' in map-executor
The new function 'addToQueue' handles adding tasks to the message queue either as a success or a failure. Another function 'unmarshal' has been introduced to unmarshal the body of an amqp.Delivery into a types.Process struct. These changes aim to enhance the handling and extraction of processes from the message queue.