Update method visibility to public in action.go and proces.go
The method "getInternalName" in both action.go and proces.go files has been changed to public method ("GetInternalName"). This allows it to be accessed from other packages as part of the API. The references to this method have also been updated in the worker actions usage and the process object setters.