Add action to add .NET project to a solution
Implemented a new dotnet action `AddProjectToSolution` in `add-to-sln.go` that allows the addition of a project to a specified solution. The action also initializes a Docker container with the required .NET SDK (version 7.0) to execute the task. The details of the solution and project, including their paths, are received as inputs. This step makes the process of managing .NET solutions more automated and efficient.