Add Docker action to install NuGet packages in .NET projects
Added a new Docker action that allows users to add NuGet packages to their .NET projects. This involves the creation of an add-nuget Docker file and a corresponding Go function to execute the action. The action will take in the project path and package name as user inputs, run the dotnet add package command in a new container, and return the container ID as an output.