Update go.sum dependencies
The changes in this commit include updates to several dependencies in go.sum. Some libraries added include cloud.google.com/go and github.com/AdamKorcz/go-118-fuzz-build, along with associated versions. It is part of an ongoing effort to keep the project dependencies up-to-date and ensure they are compatible with each