Change MongoDB collection name and connection string
The MongoDB collection name has been changed from "pipelines" to "process". Also, the connection string has been updated to include both the database user and password for better security and data protection. This adjustment allows the client to authenticate with the MongoDB server during the initiation of the connection process.