Refactor button in editor.html and enhance code update functionality
The form element container for the execute button in the editor.html file has been removed, simplifying the markup. The code update functionality has been improved with the addition of event handlers to capture changes and update accordingly. The execute button has also been equipped with a 'click' event that makes an AJAX call to '/execute', passing the updated window.code as the data payload.