Add Docker container removal functionality
The commit adds functionality to remove Docker containers. It specifically introduces a new action "docker-remove" in the action map, which results in removal of Docker container with the specified container ID. Moreover, a 'RemoveContainer' function has been added to the DockerContainer.go file, allowing the removal of Docker containers from within the application.